+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.0.0
+
+* Breaking change! Introduced `scraping.ScrapeEntity` struct which is now a part of the `NewScraper` constructor. This change allows to define additional news sources without adjusting `scraping` funcs
+* Added `scraping.PickRandomUserAgent` func to randomize user agent for each run
+* Introduced `database.CheckIfRecordWithUrlExists` func to avoid scraping news that already exist in DB
+* Added a new source
+* Reduced the value of newsAgeDays to 2 days instead of 14
+
 # 1.2.0
 
 * Improved the project structure for better readability: separated structs and functions